layers/medupdates_utilities/python/requirements.txt
sqlalchemy
psycopg2-binary
validators
transitions
boto3==1.24.56
pytz 